git.GitBranch: support checking out tags
Checking out a git repo is a last resort, so I always like to pin them to tags when I do have to have one on a host.
`git describe` is considered plumbing and thus format stable. This behavior was introduced in 2005, so there are no
real concerns that a host will be running an existing version of git that doesn't support the command or the `--all`
flag.