packaging: new utility functions
`parse_package_version` splits a package specification into the package name and the specified version

`ensure_single_package` is for package managers that allow only a single version of a package to be installed at once.

`ensure_multi_packages` is for package managers that allow multiple versions of the same package to be installed (e.g.
asdf, rubygems, etc)

`ensure_versionless_packages` is for package managers that don't have a firm concept of versions, such as asdf's plugins